Implement Swipeable in the Board component. Make individual columns vertically scrollable. Allow horizontal scroll for now.
So far this is somehow in a stable state. Handling scrolling, swiping
and tasks drag&drop in the same time is not straight forward.
For now scrolling is enabled so that columns scroll vertically
independent on one another and horizontal scroll for the whole board as
well. Dragging tasks between columns work fine - z-index is preserved
and autoscrolling while dragging is also in place.
TODO: swiping individual tasks options menu also scrolls horizontaly.
That needs a fix.